# Use properties instead of public fields for class types