--- a/src/Esprima/Ast/BlockStatement.cs
+++ b/src/Esprima/Ast/BlockStatement.cs
@@ -4,7 +4,7 @@ namespace Esprima.Ast
 {
     public class BlockStatement : Statement
     {
-        public readonly List<StatementListItem> Body;
+        public List<StatementListItem> Body { get; }
 
         public BlockStatement(List<StatementListItem> body)
         {