--- a/src/Esprima/Ast/BreakStatement.cs
+++ b/src/Esprima/Ast/BreakStatement.cs
@@ -2,7 +2,7 @@ namespace Esprima.Ast
 {
     public class BreakStatement : Statement
     {
-        public readonly Identifier Label;
+        public Identifier Label { get; }
 
         public BreakStatement(Identifier label)
         {