--- a/src/Esprima/Ast/CallExpression.cs
+++ b/src/Esprima/Ast/CallExpression.cs
@@ -3,20 +3,19 @@ using Newtonsoft.Json;
 
 namespace Esprima.Ast
 {
-    public class CallExpression : Node,
-        Expression
+    public class CallExpression : Node, Expression
     {
-        public readonly Expression Callee;
-        public readonly List<ArgumentListElement> Arguments;
+        public Expression Callee { get; }
+        public List<ArgumentListElement> Arguments { get; }
 
         [JsonIgnore]
-        public bool Cached;
+        public bool Cached { get; set; }
 
         [JsonIgnore]
-        public bool CanBeCached = true;
+        public bool CanBeCached { get; set;} = true;
 
         [JsonIgnore]
-        public object CachedArguments;
+        public object CachedArguments { get; set; }
 
         public CallExpression(Expression callee, List<ArgumentListElement> args)
         {