--- a/src/Esprima/Ast/CatchClause.cs
+++ b/src/Esprima/Ast/CatchClause.cs
@@ -2,8 +2,8 @@ namespace Esprima.Ast
 {
     public class CatchClause : Statement
     {
-        public readonly ArrayPatternElement Param; // BindingIdentifier | BindingPattern;
-        public readonly BlockStatement Body;
+        public ArrayPatternElement Param { get; } // BindingIdentifier | BindingPattern;
+        public BlockStatement Body { get; }
 
         public CatchClause(ArrayPatternElement param, BlockStatement body)
         {