--- a/src/Esprima/Ast/ClassExpression.cs
+++ b/src/Esprima/Ast/ClassExpression.cs
@@ -2,9 +2,9 @@
 {
     public class ClassExpression : Node, Expression
     {
-        public readonly Identifier Id;
-        public readonly PropertyKey SuperClass;
-        public readonly ClassBody Body;
+        public Identifier Id { get; }
+        public PropertyKey SuperClass { get; }
+        public ClassBody Body { get; }
 
         public ClassExpression(Identifier id, PropertyKey superClass, ClassBody body)
         {