--- a/src/Esprima/Ast/Directive.cs
+++ b/src/Esprima/Ast/Directive.cs
@@ -5,7 +5,7 @@ namespace Esprima.Ast
     public class Directive : ExpressionStatement
     {
         [JsonProperty("directive")]
-        public readonly string Directiv;
+        public string Directiv { get; }
 
         public Directive(Expression expression, string directive)
             :base(expression)