--- a/src/Esprima/Ast/ForStatement.cs
+++ b/src/Esprima/Ast/ForStatement.cs
@@ -3,10 +3,10 @@ namespace Esprima.Ast
     public class ForStatement : Statement
     {
         // can be a Statement (var i) or an Expression (i=0)
-        public readonly INode Init;
-        public readonly Expression Test;
-        public readonly Expression Update;
-        public readonly Statement Body;
+        public INode Init { get; }
+        public Expression Test { get; }
+        public Expression Update { get; }
+        public Statement Body { get; }
 
         public ForStatement(INode init, Expression test, Expression update, Statement body)
         {