--- a/src/Esprima/Ast/ObjectPattern.cs
+++ b/src/Esprima/Ast/ObjectPattern.cs
@@ -4,7 +4,7 @@ namespace Esprima.Ast
 {
     public class ObjectPattern : Node, BindingPattern
     {
-        public readonly List<Property> Properties;
+        public List<Property> Properties { get; }
 
         public ObjectPattern(List<Property> properties)
         {