--- a/src/Esprima/ParserException.cs
+++ b/src/Esprima/ParserException.cs
@@ -4,10 +4,10 @@ namespace Esprima
 {
     public class ParserException : Exception
     {
-        public int Column;
-        public string Description;
-        public int Index;
-        public int LineNumber;
+        public int Column { get; set;}
+        public string Description { get; set;}
+        public int Index { get; set; }
+        public int LineNumber { get; set;}
 
         public ParserException(string message) : base(message)
         {