--- a/src/Esprima/SourceType.cs
+++ b/src/Esprima/SourceType.cs
@@ -1,5 +1,8 @@
-﻿public enum SourceType
+﻿namespace Esprima
 {
-    Module,
-    Script
+    public enum SourceType
+    {
+        Module,
+        Script
+    }
 }