… src/Esprima/JavascriptParser.cs
… src/Esprima/State.cs
… src/Esprima/Token.cs