API cleanup, now more idiomatic with classes having properties instead of public fields. Performance does not seem to suffer noticeably. We are still pretty much within the performance boost numbers we got from the bigger PR which improved performance.

Changing from field to property is binary breaking change (binding redirect won't work) so if you want to do this, probably better to be done before 1.0.

### Before

|       Method |            FileName |      Mean |     Error |    StdDev |     Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|------------- |-------------------- |----------:|----------:|----------:|----------:|---------:|---------:|----------:|
| **ParseProgram** |       **angular-1.2.5** | **32.024 ms** | **0.2408 ms** | **0.2252 ms** | **1937.5000** | **750.0000** | **250.0000** |  **10.15 MB** |
| **ParseProgram** |      **backbone-1.1.0** |  **2.661 ms** | **0.0200 ms** | **0.0167 ms** |  **285.1563** | **140.6250** |        **-** |   **1.56 MB** |
| **ParseProgram** |        **jquery-1.9.1** | **22.935 ms** | **0.3868 ms** | **0.3429 ms** | **1593.7500** | **656.2500** | **125.0000** |   **8.64 MB** |
| **ParseProgram** | **jquery.mobile-1.4.2** | **36.748 ms** | **0.1733 ms** | **0.1621 ms** | **2437.5000** | **937.5000** | **250.0000** |  **13.21 MB** |
| **ParseProgram** |      **mootools-1.4.5** | **18.630 ms** | **0.1480 ms** | **0.1312 ms** | **1343.7500** | **593.7500** | **125.0000** |   **7.12 MB** |
| **ParseProgram** |    **underscore-1.5.2** |  **2.093 ms** | **0.0092 ms** | **0.0086 ms** |  **292.9688** |  **58.5938** |        **-** |   **1.35 MB** |
| **ParseProgram** |          **yui-3.12.0** | **14.400 ms** | **0.0264 ms** | **0.0234 ms** | **1093.7500** | **546.8750** |        **-** |   **6.43 MB** |

### After

|       Method |            FileName |      Mean |     Error |    StdDev |     Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|------------- |-------------------- |----------:|----------:|----------:|----------:|---------:|---------:|----------:|
| **ParseProgram** |       **angular-1.2.5** | **32.682 ms** | **0.2284 ms** | **0.2137 ms** | **1937.5000** | **750.0000** | **250.0000** |  **10.15 MB** |
| **ParseProgram** |      **backbone-1.1.0** |  **2.808 ms** | **0.0031 ms** | **0.0026 ms** |  **285.1563** | **140.6250** |        **-** |   **1.56 MB** |
| **ParseProgram** |        **jquery-1.9.1** | **23.411 ms** | **0.0719 ms** | **0.0672 ms** | **1593.7500** | **656.2500** | **125.0000** |   **8.64 MB** |
| **ParseProgram** | **jquery.mobile-1.4.2** | **38.243 ms** | **0.1129 ms** | **0.1056 ms** | **2437.5000** | **937.5000** | **250.0000** |  **13.21 MB** |
| **ParseProgram** |      **mootools-1.4.5** | **19.292 ms** | **0.0622 ms** | **0.0552 ms** | **1343.7500** | **593.7500** | **125.0000** |   **7.12 MB** |
| **ParseProgram** |    **underscore-1.5.2** |  **2.254 ms** | **0.0030 ms** | **0.0028 ms** |  **292.9688** |  **58.5938** |        **-** |   **1.35 MB** |
| **ParseProgram** |          **yui-3.12.0** | **15.767 ms** | **0.0402 ms** | **0.0376 ms** | **1093.7500** | **546.8750** |        **-** |   **6.43 MB** |