@lahma 2017-12-27
Closing in favor of #16 which is less ambitious and has no performance impact.